--- a/Cargo.toml
+++ b/Cargo.toml
@@ -3,5 +3,6 @@ resolver = "2"
 members = [
     "valuable",
     "valuable-derive",
+    "valuable-json",
     "tests",
 ]