--- a/valuable-json/Cargo.toml
+++ b/valuable-json/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "valuable-json"
+version = "0.1.0"
+authors = ["Taiki Endo <te316e89@gmail.com>"]
+edition = "2018"
+license = "MIT"
+description = "Values"
+
+[features]
+default = ["std"]
+
+# NOTE: Disabling `std` feature is not supported yet.
+std = ["valuable/std"]
+
+[dependencies]
+valuable = { version = "0.1", path = "../valuable", default-features = false }
+
+[dev-dependencies]
+valuable = { version = "0.1", path = "../valuable", features = ["derive"] }
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"