@taiki-e 2021-05-20
Currently, this is almost the same as the JSON serializer example that I previously wrote. Hopefully, I can get the rest of the implementation work done over the weekend and into next week.