@taiki-e 2021-06-24
Thanks @KodrAus! That benchmark is very useful.

---

It seems that one of the problems is that the `escape` function is not inlined. Inlining the `escape` function would greatly improve performance.

Before (2374aa8073b013c09e0ce6398502c53ff24744fc):

```
test twitter_valuable   ... bench:   1,890,449 ns/iter (+/- 121,674)
```
After (07a320cb87693ead56754c32db82bc377cbd5360):
```
test twitter_valuable   ... bench:   1,010,623 ns/iter (+/- 198,776)
```

However, it's still much slower than 2665f14e9f3583f9d78ad1894c934d036d4fa492.

And, escaping seems to account for around 40% of the total. The following are the results when escaping is disabled (f7e335ce2c71afbe2c3a4eb420ffebc61540e36b).

```
test twitter_valuable   ... bench:     634,698 ns/iter (+/- 53,980)
```

Also, the following two changes will also improve performance:

- use serde_json's escaping implementation (https://github.com/serde-rs/json/pull/89).
- remove utf8 check

```
test twitter_valuable     ... bench:     812,378 ns/iter (+/- 29,584)
```

One of the rest of the problems is that the style options are checked frequently. Removing them will look like the following:

```
test twitter_erased_serde ... bench:     609,975 ns/iter (+/- 92,778)
test twitter_miniserde    ... bench:     988,489 ns/iter (+/- 62,401)
test twitter_serde        ... bench:     419,275 ns/iter (+/- 43,249)
test twitter_sval         ... bench:     663,869 ns/iter (+/- 32,564)
test twitter_valuable     ... bench:     685,965 ns/iter (+/- 38,047)
```